requirements.txt
asgiref==3.9.1
asttokens==3.0.0
babel==2.17.0
click==8.3.0
colorama==0.4.6
decorator==5.2.1
diff-match-patch==20241021
Django==5.2.6
django-admin-kubi==1.1.1
django-appconf==1.1.0
django-compressor==4.5.1
django-import-export==4.3.10
django-money==3.5.4
django-otp==1.6.3
et_xmlfile==2.0.0
executing==2.2.1
Faker==37.6.0
gunicorn==23.0.0
h11==0.16.0
iniconfig==2.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
load-dotenv==0.1.0
matplotlib-inline==0.1.7
openpyxl==3.1.5
packaging==25.0
parso==0.8.5
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-moneyed==3.0
Pygments==2.19.2
pyotp==2.9.0
pytest==8.4.2
python-dotenv==1.1.1
rcssmin==1.1.2
rjsmin==1.2.2
setuptools==80.9.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
wcwidth==0.2.14
whitenoise==6.8.2
wkhtmltopdf==0.2
